NW: dict vs generator?
