+++ pit2ya/user_start.py
@@ -28,75 +28,36 @@ from dataclasses import dataclass
 #     for i in range(10):
 #         print(pq.get())
 
-async def get_data():
-    from pickle import dump, load
-    from toggl.api import TimeEntry
-    from pendulum import now
-    from os import getenv, path
-    filepath = getenv('PIT2YA_DIRPATH') or getenv('XDG_DATA_HOME') + '/pit2ya/timers.pykle'
-    timers = {}
-    if path.isfile(filepath):
-        with open(filepath, 'rb') as rf:
-            timers = load(rf)
-    else:
-        print("Pit2ya data not found... please try again after data has been processed")
-        # from os import mkdir
-        # print('config file not found... loading past month of toggl data')
-        # entries = TimeEntry.objects.all_from_reports(start=now().subtract(months=1), stop=now())
-        # print('    processing time entries...')
-        # if not path.isdir(filepath[:filepath.rfind('/')]):
-        #     mkdir(filepath[:filepath.rfind('/')])
-        # seen = set()
-        # for i,e in enumerate(entries):
-        #     if e.description not in seen:
-        #         seen.add(e.description)
-        #         timers[e.description] = { 'pid': int(e.pid or -1) }
-        #         # writer.writerow((e.description, e.pid or -1))
-        #     if i % 10 == 0:
-        #         print(f'processed {i} entries', end='\r')
-        # with open(filepath, 'wb') as wf:
-        #     dump(timers, wf)
-    return timers
+# async def get_data():
+#     from pickle import dump, load
+#     from toggl.api import TimeEntry
+#     from pendulum import now
+#     from os import getenv, path
+#     filepath = getenv('PIT2YA_DIRPATH') or getenv('XDG_DATA_HOME') + '/pit2ya/timers.pykle'
+#     timers = {}
+#     if path.isfile(filepath):
+#         with open(filepath, 'rb') as rf:
+#             timers = load(rf)
+#     else:
+#         print("Pit2ya data not found... please try again after data has been processed")
+#         # from os import mkdir
+#         # print('config file not found... loading past month of toggl data')
+#         # entries = TimeEntry.objects.all_from_reports(start=now().subtract(months=1), stop=now())
+#         # print('    processing time entries...')
+#         # if not path.isdir(filepath[:filepath.rfind('/')]):
+#         #     mkdir(filepath[:filepath.rfind('/')])
+#         # seen = set()
+#         # for i,e in enumerate(entries):
+#         #     if e.description not in seen:
+#         #         seen.add(e.description)
+#         #         timers[e.description] = { 'pid': int(e.pid or -1) }
+#         #         # writer.writerow((e.description, e.pid or -1))
+#         #     if i % 10 == 0:
+#         #         print(f'processed {i} entries', end='\r')
+#         # with open(filepath, 'wb') as wf:
+#         #     dump(timers, wf)
+#     return timers
 
-def begin_timer_raw(desc, pid):
-    from toggl.api import TimeEntry
-    from pendulum import now
-    if pid >= 0:
-        TimeEntry.start_and_save(start=now(), description=desc, pid=pid).save()
-    else:
-        TimeEntry.start_and_save(start=now(), pid=pid).save()
-
-async def user_start():
-    from iterfzf import iterfzf
-    timers = await get_data()
-    query, desc = iterfzf(timers.keys(), print_query=True, extended=True)
-
-    if desc:
-        begin_timer_raw(desc, timers[desc]['pid'])
-    else:
-        pass    # TODO: collect project information, allow creating new time entries
-        # project = input(f"Creating new time entry '{query}'... what project? ")
-async def user_modify():
-    from iterfzf import iterfzf
-    timers = await get_data()
-    query, desc = iterfzf(timers.keys(), print_query=True, extended=True)
-
-    from toggl.api import TimeEntry
-
-    cur = TimeEntry.objects.current()
-    if cur is None:
-        print('No current running timer!')
-        begin_timer_raw(desc, timers[desc]['pid'])
-    elif desc:
-        setattr(cur, 'description', desc)
-        setattr(cur, 'project', timers[desc]['pid'])
-        cur.save()
-
-def entry_start():
-    asyncio.run(user_start())
-
-def entry_modify():
-    asyncio.run(user_modify())
 
 if __name__ == '__main__':
     entry_start()
